requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
python-multipart
aiofiles
requests
Jinja2
PyMySQL
SQLAlchemy
python-dotenv
